\name{etfdata}
\docType{data}
\alias{etfdata}
\title{15 Exchange Traded Funds (ETFs)}
\description{
The xts dataset consists of the adjusted daily closing prices of 15 
iShare ETFs for the period 2003-05-28 to 2012-06-01 (2272 periods) representing 
a selection of US style and international equity benchmark. The iShare 
series are IWF (Russell 1000 Growth Index), IWD (Russell 1000 Value Index),
IWO (Russell 2000 Growth Index), IWN (Russell 2000 Value Index),
EEM (MSCI Emerging Markets Index), TLT (Barclays 20+ Year T-Bond),
EWC (MSCI Canada Index), EWA (MSCI Australia Index), EWJ (MSCI Japan Index),
EWG (MSCI Germany Index), EWL (MSCI Switzerland Index), EWQ (MSCI France Index),
EWU (MSCI UK Index), EPP (MSCI Pacific ex-Japan), EZA (MSCI South Africa Index).
}
\usage{data(etfdata)}
\format{An xts matrix containing 2272x15 observations.}
\source{Yahoo Finance}
\keyword{datasets}
